@@ -1,6 +1,6 @@
 module github.com/darkweak/souin/plugins/gin
 
-go 1.16
+go 1.18
 
 require (
 	github.com/darkweak/souin v1.6.18
