@@ -2,7 +2,7 @@ package chi
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -96,7 +96,7 @@ func Test_SouinChiPlugin_Middleware_APIHandle(t *testing.T) {
 	if res.Result().Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must be in JSON.")
 	}
-	b, _ := ioutil.ReadAll(res.Result().Body)
+	b, _ := io.ReadAll(res.Result().Body)
 	res.Result().Body.Close()
 	if string(b) != "[]" {
 		t.Error("The response body must be an empty array because no request has been stored")
@@ -106,7 +106,7 @@ func Test_SouinChiPlugin_Middleware_APIHandle(t *testing.T) {
 	if res2.Result().Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must be in JSON.")
 	}
-	b, _ = ioutil.ReadAll(res2.Result().Body)
+	b, _ = io.ReadAll(res2.Result().Body)
 	res2.Result().Body.Close()
 	var payload []string
 	_ = json.Unmarshal(b, &payload)
