@@ -2,7 +2,8 @@ package souin
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -51,6 +52,7 @@ func Test_SouinEchoPlugin_Process(t *testing.T) {
 		t.Error("No error must be thrown on the second request if everything is good.")
 	}
 	if res2.Result().Header.Get("Cache-Status") != "Souin; hit; ttl=4" {
+		fmt.Println(res2.Result().Header.Get("Cache-Status"))
 		t.Error("The response must contain a Cache-Status header with the hit and ttl directives.")
 	}
 	if res2.Result().Header.Get("Age") != "1" {
@@ -112,7 +114,7 @@ func Test_SouinEchoPlugin_Process_APIHandle(t *testing.T) {
 	if res.Result().Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must contain be in JSON.")
 	}
-	b, _ := ioutil.ReadAll(res.Result().Body)
+	b, _ := io.ReadAll(res.Result().Body)
 	defer res.Result().Body.Close()
 	if string(b) != "[]" {
 		t.Error("The response body must be an empty array because no request has been stored")
@@ -126,7 +128,7 @@ func Test_SouinEchoPlugin_Process_APIHandle(t *testing.T) {
 	if res.Result().Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must contain be in JSON.")
 	}
-	b, _ = ioutil.ReadAll(res2.Result().Body)
+	b, _ = io.ReadAll(res2.Result().Body)
 	defer res.Result().Body.Close()
 	var payload []string
 	_ = json.Unmarshal(b, &payload)
