@@ -3,7 +3,7 @@ package kratos
 import (
 	"bytes"
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -106,14 +106,14 @@ func Test_HttpcacheKratosPlugin_NewHTTPCacheFilter_Mutation(t *testing.T) {
 	time.Sleep(config.DefaultCache.GetTTL())
 	handler := NewHTTPCacheFilter(config)(nextFilter)
 	req, res, res2 := prepare("/handled")
-	req.Body = ioutil.NopCloser(bytes.NewBuffer([]byte(`{"query":"mutation":{something mutated}}`)))
+	req.Body = io.NopCloser(bytes.NewBuffer([]byte(`{"query":"mutation":{something mutated}}`)))
 	handler.ServeHTTP(res, req)
 	rs := res.Result()
 	rs.Body.Close()
 	if rs.Header.Get("Cache-Status") != "Souin; fwd=uri-miss" {
 		t.Error("The response must contain a Cache-Status header without the stored directive and with the uri-miss only.")
 	}
-	req.Body = ioutil.NopCloser(bytes.NewBuffer([]byte(`{"query":"mutation":{something mutated}}`)))
+	req.Body = io.NopCloser(bytes.NewBuffer([]byte(`{"query":"mutation":{something mutated}}`)))
 	handler.ServeHTTP(res2, req)
 	rs = res2.Result()
 	rs.Body.Close()
@@ -135,7 +135,7 @@ func Test_HttpcacheKratosPlugin_NewHTTPCacheFilter_API(t *testing.T) {
 	if rs.Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must contain be in JSON.")
 	}
-	b, _ := ioutil.ReadAll(rs.Body)
+	b, _ := io.ReadAll(rs.Body)
 	res.Result().Body.Close()
 	if string(b) != "[]" {
 		t.Error("The response body must be an empty array because no request has been stored")
@@ -154,7 +154,7 @@ func Test_HttpcacheKratosPlugin_NewHTTPCacheFilter_API(t *testing.T) {
 	if rs.Header.Get("Content-Type") != "application/json" {
 		t.Error("The response must contain be in JSON.")
 	}
-	b, _ = ioutil.ReadAll(rs.Body)
+	b, _ = io.ReadAll(rs.Body)
 	rs.Body.Close()
 	var payload []string
 	_ = json.Unmarshal(b, &payload)
