@@ -15,7 +15,7 @@ import (
 func mockCloudflareProvider() *CloudflareSurrogateStorage {
 	ass := &CloudflareSurrogateStorage{
 		baseStorage: &baseStorage{
-			Storage:    make(map[string]string),
+			Storage:    &sync.Map{},
 			Keys:       make(map[string]configurationtypes.SurrogateKeys),
 			keysRegexp: make(map[string]keysRegexpInner),
 			dynamic:    true,
