@@ -13,7 +13,7 @@ import (
 func mockSouinProvider() *SouinSurrogateStorage {
 	sss := &SouinSurrogateStorage{
 		baseStorage: &baseStorage{
-			Storage:    make(map[string]string),
+			Storage:    &sync.Map{},
 			Keys:       make(map[string]configurationtypes.SurrogateKeys),
 			keysRegexp: make(map[string]keysRegexpInner),
 			dynamic:    true,
