@@ -106,6 +106,7 @@ func (s *SouinChiMiddleware) Handle(next http.Handler) http.Handler {
 			Response: &http.Response{},
 			Buf:      s.bufPool.Get().(*bytes.Buffer),
 			Rw:       rw,
+			Req:      req,
 		}
 		req = s.Retriever.GetContext().SetContext(req)
 		getterCtx := getterContext{next.ServeHTTP, customWriter, req}
@@ -124,11 +125,8 @@ func (s *SouinChiMiddleware) Handle(next http.Handler) http.Handler {
 			combo.next.ServeHTTP(customWriter, r)
 
 			combo.req.Response = customWriter.Response
-			if combo.req.Response, e = s.Retriever.GetTransport().(*rfc.VaryTransport).UpdateCacheEventually(combo.req); e != nil {
-				return e
-			}
+			combo.req.Response, e = s.Retriever.GetTransport().(*rfc.VaryTransport).UpdateCacheEventually(combo.req)
 
-			_, _ = customWriter.Send()
 			return e
 		})
 	})
