@@ -1,7 +1,6 @@
 package fiber
 
 import (
-	"bytes"
 	"net/http"
 
 	"github.com/darkweak/souin/plugins"
@@ -12,11 +11,22 @@ type (
 	fiberWriterDecorator struct {
 		*plugins.CustomWriter
 	}
+	nopWriter struct {
+		h http.Header
+		*fiber.Ctx
+	}
 	fiberWriter struct {
 		h http.Header
 		*fiber.Ctx
 	}
 )
+
+func (f *nopWriter) Header() http.Header {
+	if f.h == nil {
+		f.h = http.Header{}
+	}
+	return f.h
+}
 
 func (f *fiberWriter) Header() http.Header {
 	if f.h == nil {
@@ -29,8 +39,18 @@ func (f *fiberWriter) WriteHeader(code int) {
 	f.Ctx.Response().Header.SetStatusCode(code)
 }
 
+func (f *nopWriter) WriteHeader(code int) {
+	f.Ctx.Response().Header.SetStatusCode(code)
+}
+
+func (f *nopWriter) Write(b []byte) (int, error) {
+	return len(b), nil
+}
+
+func (f *fiberWriter) Write(b []byte) (int, error) {
+	return f.Ctx.Write(b)
+}
+
 func (f *fiberWriterDecorator) Send() (int, error) {
-	b := new(bytes.Buffer)
-	_, _ = b.ReadFrom(f.CustomWriter.Response.Body)
-	return b.Len(), nil
+	return f.CustomWriter.Send()
 }
