@@ -100,6 +100,7 @@ func (s *SouinGoZeroMiddleware) Handle(next http.HandlerFunc) http.HandlerFunc {
 			Response: &http.Response{},
 			Buf:      s.bufPool.Get().(*bytes.Buffer),
 			Rw:       rw,
+			Req:      req,
 		}
 		req = s.Retriever.GetContext().SetContext(req)
 		getterCtx := getterContext{next, customWriter, req}
@@ -118,11 +119,8 @@ func (s *SouinGoZeroMiddleware) Handle(next http.HandlerFunc) http.HandlerFunc {
 			combo.next(customWriter, r)
 
 			combo.req.Response = customWriter.Response
-			if combo.req.Response, e = s.Retriever.GetTransport().(*rfc.VaryTransport).UpdateCacheEventually(combo.req); e != nil {
-				return e
-			}
+			combo.req.Response, e = s.Retriever.GetTransport().(*rfc.VaryTransport).UpdateCacheEventually(combo.req)
 
-			_, _ = customWriter.Send()
 			return e
 		})
 	}
