@@ -2,8 +2,6 @@ package rfc
 
 import (
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net/http"
 
 	"github.com/darkweak/souin/context"
@@ -30,19 +28,11 @@ func validateVary(req *http.Request, resp *http.Response, key string, t *VaryTra
 		}
 		resp.Header.Set("Cache-Status", fmt.Sprintf("%s; fwd=uri-miss; stored", req.Context().Value(context.CacheName)))
 		resp.Header.Del("Age")
-		// Delay caching until EOF is reached.
-		resp.Body = &cachingReadCloser{
-			R: resp.Body,
-			OnEOF: func(r io.Reader) {
-				re := *resp
-				re.Body = ioutil.NopCloser(r)
-				_ = t.SurrogateStorage.Store(&re, cacheKey)
-				t.SetCache(cacheKey, &re)
-				go func() {
-					t.CoalescingLayerStorage.Delete(cacheKey)
-				}()
-			},
-		}
+		_ = t.SurrogateStorage.Store(resp, cacheKey)
+		t.SetCache(cacheKey, resp)
+		go func() {
+			t.CoalescingLayerStorage.Delete(cacheKey)
+		}()
 		return true
 	}
 
